(REF): Minor annotators refactoring
