+++ src/main/kotlin/org/rust/ide/annotator/RustItemsAnnotator.kt
@@ -1,18 +1,16 @@
 package org.rust.ide.annotator
 
-import com.intellij.codeInspection.LocalQuickFixAndIntentionActionOnPsiElement
 import com.intellij.lang.annotation.AnnotationHolder
 import com.intellij.lang.annotation.Annotator
-import com.intellij.openapi.editor.Editor
-import com.intellij.openapi.project.Project
 import com.intellij.openapi.util.TextRange
 import com.intellij.psi.PsiElement
-import com.intellij.psi.PsiFile
 import org.rust.cargo.project.workspace.cargoProject
-import org.rust.ide.actions.RustExpandModuleAction
-import org.rust.lang.core.psi.*
-import org.rust.lang.core.psi.impl.RustFile
-import org.rust.lang.core.psi.impl.mixin.getOrCreateModuleFile
+import org.rust.ide.annotator.fixes.AddModuleFileFix
+import org.rust.ide.annotator.fixes.ImplementMethodsFix
+import org.rust.lang.core.psi.RustElementVisitor
+import org.rust.lang.core.psi.RustImplItemElement
+import org.rust.lang.core.psi.RustModDeclItemElement
+import org.rust.lang.core.psi.containingMod
 import org.rust.lang.core.psi.impl.mixin.isAbstract
 import org.rust.lang.core.psi.impl.mixin.isLocal
 import org.rust.lang.core.psi.impl.mixin.pathAttribute
@@ -46,14 +44,14 @@ class RustItemsAnnotator : Annotator {
             // sure that a mod is not a directory owner.
             if (modDecl.module?.cargoProject != null) {
                 holder.createErrorAnnotation(modDecl, "Cannot declare a new module at this location")
-                    .registerFix(AddModuleFile(modDecl, expandModuleFirst = true))
+                    .registerFix(AddModuleFileFix(modDecl, expandModuleFirst = true))
             }
             return
         }
 
         if (modDecl.reference.resolve() == null) {
             holder.createErrorAnnotation(modDecl, "Unresolved module")
-                .registerFix(AddModuleFile(modDecl, expandModuleFirst = false))
+                .registerFix(AddModuleFileFix(modDecl, expandModuleFirst = false))
         }
     }
 
@@ -74,7 +72,7 @@ class RustItemsAnnotator : Annotator {
 
             holder.createErrorAnnotation(implHeaderTextRange,
                 "Not all trait items implemented, missing: `${notImplemented.first()}`"
-            ).registerFix(ImplementMethods(impl, toImplement))
+            ).registerFix(ImplementMethodsFix(impl, toImplement))
 
         }
 
@@ -84,58 +82,4 @@ class RustItemsAnnotator : Annotator {
                 "Method is not a member of trait `${trait.name}`")
         }
     }
-}
-
-private class AddModuleFile(
-    modDecl: RustModDeclItemElement,
-    private val expandModuleFirst: Boolean
-) : LocalQuickFixAndIntentionActionOnPsiElement(modDecl) {
-    override fun getText(): String = "Create module file"
-
-    override fun getFamilyName(): String = text
-
-    override fun invoke(
-        project: Project,
-        file: PsiFile,
-        editor: Editor?,
-        startElement: PsiElement,
-        endElement: PsiElement
-    ) {
-        val modDecl = startElement as RustModDeclItemElement
-        if (expandModuleFirst) {
-            val containingFile = modDecl.containingFile as RustFile
-            RustExpandModuleAction.expandModule(containingFile)
-        }
-        modDecl.getOrCreateModuleFile()?.navigate(true)
-    }
-
-}
-
-private class ImplementMethods(
-    implBody: RustImplItemElement,
-    private val methods: List<RustFunctionElement>
-) : LocalQuickFixAndIntentionActionOnPsiElement(implBody) {
-    init {
-        check(methods.isNotEmpty())
-    }
-
-    override fun getText(): String = "Implement methods"
-
-    override fun getFamilyName(): String = text
-
-    override fun invoke(
-        project: Project,
-        file: PsiFile,
-        editor: Editor?,
-        startElement: PsiElement,
-        endElement: PsiElement
-    ) {
-        val impl = (startElement as RustImplItemElement)
-        val templateImpl = RustPsiFactory(project).createTraitImplItem(methods)
-        val lastMethodOrBrace = impl.functionList.lastOrNull() ?: impl.lbrace ?: return
-        val firstToAdd = templateImpl.functionList.first()
-        val lastToAdd = templateImpl.functionList.last()
-        impl.addRangeAfter(firstToAdd, lastToAdd, lastMethodOrBrace)
-    }
-
 }
