--- src/test/resources/org/rust/ide/annotator/fixtures/expressions/struct_expr_quick_fix2.rs
@@ -1,13 +1,0 @@
-struct S {
-    a: i32,
-    b: i32,
-    c: i32,
-    d: i32
-}
-
-fn main() {
-    let _ = S {
-        a: 92,
-        c: 92<caret>
-    };
-}
