--- src/test/resources/org/rust/ide/annotator/fixtures/expressions/struct_expr_quick_fix2_after.rs
@@ -1,15 +1,0 @@
-struct S {
-    a: i32,
-    b: i32,
-    c: i32,
-    d: i32
-}
-
-fn main() {
-    let _ = S {
-        b: (),
-        d: (),
-        a: 92,
-        c: 92
-    };
-}
