--- src/test/resources/org/rust/ide/annotator/fixtures/literals/literal_suffixes.rs
@@ -1,8 +1,0 @@
-fn main() {
-    let lit1 = <error descr="string literal with a suffix is invalid">"test"u8</error>;
-    let lit2 = <error descr="char literal with a suffix is invalid">'c'u8</error>;
-    let lit3 = <error descr="invalid suffix 'u8' for float literal; the suffix must be one of: 'f32', 'f64'">1.2u8</error>;
-    let lit4 = <error descr="invalid suffix 'f34' for float literal; the suffix must be one of: 'f32', 'f64'">1f34</error>;
-    let lit4 = <error descr="invalid suffix 'u96' for integer literal; the suffix must be one of: 'u8', 'i8', 'u16', 'i16', 'u32', 'i32', 'u64', 'i64', 'isize', 'usize'">1u96</error>;
-    let lit5 = 1e30f64;
-}
