--- src/test/resources/org/rust/ide/annotator/fixtures/literals/literal_unclosed_quotes.rs
@@ -1,4 +1,0 @@
-fn main() {
-    let ch1 = <error descr="unclosed char literal">'1</error>;
-    let ch2 = <error descr="unclosed byte literal">b'1</error>;
-}
